eventbuilder: fix the -Werror warnings CI trips on
The branch had never been through CI, so nothing had ever compiled it
with -Werror -Wall -Wextra -Wundef. Sweeping all 39 branch-modified
files with those flags locally found 17 error sites; CI itself only ever
reported the first file, because the build aborts there.

Substantive one first:

  EventBuilder/EventPrefilter/TimeAlignment/TimeCoincidence factories
  declared the JOmniFactory callback as Process(int64_t, uint64_t), but
  the base declares it virtual as Process(int32_t, uint64_t), so ours
  hid the base overloads instead of overriding them
  (-Woverloaded-virtual). Not a dispatch bug -- JOmniFactory calls it as
  static_cast<AlgoT*>(this)->Process(...), i.e. statically through the
  derived type, so our body did run -- but it silently opted out of the
  virtual interface. Use int32_t and match every other EICrecon factory.

  eventbuilder.cc left m_default_cfg out of seven TypedWiring designated
  initialisers (-Wmissing-designated-field-initializers). Value-
  initialised either way, so this is a no-op; spelled out to match the
  member order in JOmniFactoryGeneratorT::TypedWiring.

The rest are unused entities, marked [[maybe_unused]] rather than
deleted so the in-progress fit/report helpers survive: fitBackground,
split, drawLines and the vestigial `table` parameter in PdfReport.cc,
pct in TriggerBenchmark_service.cc, m_app in ONNXRuntime_service.h, and
t0sigma/phys in eventbuilder.cc, which document the weight indices they
read. One genuinely dead local (`blocks`) is removed.

EnableOrdering(true) is now guarded by a compile-time detection trait
rather than called unconditionally: the API is absent from the older
JANA2 in the eic-shell *-stable images, which broke both stable builds.
Detecting it keeps ordering everywhere it exists instead of pinning a
version -- losing it would not affect the totals (order-free sums) but
would make the interim "after N frames" reports cover a different N
frames per thread count.

All 39 files now compile clean under the CI flags; eventbuilder tests
still 41 passed / 2 skipped; the trigger benchmark reproduces the
reference report exactly.

+++ src/factories/eventbuilder/EventBuilder_factory.h
@@ -591,7 +591,7 @@ struct EventBuilder_factory : public JOmniFactory<EventBuilder_factory> {
     }
   };
 
-  void Process(int64_t run_number, uint64_t event_number) {
+  void Process(int32_t run_number, uint64_t event_number) {
     // Topology grid dimensions (runtime; two staggered grids of ntheta x nphi).
     const int ntheta = std::max(1, m_theta_bins);
     const int nphi   = std::max(1, m_phi_bins);
